telegram-bot/src/main/java/pro/sky/telegrambot/service/UsersService.java
package pro.sky.telegrambot.service;


import org.springframework.stereotype.Service;
import pro.sky.telegrambot.model.Pet;
import pro.sky.telegrambot.model.User;
import pro.sky.telegrambot.repository.UsersRepository;

import java.util.Optional;

@Service
public class UsersService {

    private final UsersRepository usersRepository;

    public UsersService(UsersRepository usersRepository) {
        this.usersRepository = usersRepository;
    }


    public User addUser(User user) {
        return usersRepository.save(user);
    }

    public Optional<User> findUserById(Long id) {
        return usersRepository.findById(id);
    }

    public User editUser(User user) {
        return usersRepository.save(user);
    }

    public void deleteUser(Long id) {
        usersRepository.deleteById(id);
    }
}



